Add animation and movement for invaders
Bugs: Invaders SHOULD bounce of edges, move closer to player, and change direction but they dont. Implementation in invader_spawner.gd and the walls in the main scene.

When changing the window size to meet the project requirements, i think it messed up some of the positioning used in the tutorial since its based on pixels. I hard-coded a random value that I think is ok. We might want to revisit that later.